+++ ItkQuillEditorBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Itk\QuillEditorBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class ItkQuillEditorBundle extends Bundle
-{
-}