+++ ItkTemplateExtensionBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Itk\TemplateExtensionBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class ItkTemplateExtensionBundle extends Bundle
+{
+}