+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Quill Editor Bundle
+# Template Extension Bundle
+
+## In develop
+
+* Renamed bundle to Template Extension Bundle.
+* Added quill-text editor.
 
 ## 1.0.0